pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sccraper-it"
version = "1.1.1"
dependencies = [
    "requests",
    "python-dotenv",
    "tqdm"
]

[project.scripts]
# This points to the main() function inside src/main.py
sccraper = "src.main:main"

[tool.setuptools.packages.find]
# This tells setuptools that the root of everything is the current directory
where = ["."]
include = ["src*"]
